+++ jimregexp.c
@@ -5,6 +5,15 @@
  */
 
 /*
+ * Stolen from musl (src/include/features.h)
+ *
+ * SPDX-FileCopyrightText: The musl Contributors
+ *
+ * SPDX-License-Identifier: MIT
+ */
+#define hidden __attribute__((__visibility__("hidden")))
+
+/*
  * Stolen from jimtcl (jim.h)
  *
  * SPDX-FileCopyrightText: The jimctl Contributors
@@ -274,7 +283,7 @@ static int str_int_len(const int *seq)
  * Beware that the optimization-preparation code in here knows about some
  * of the structure of the compiled regexp.
  */
-int jim_regcomp(regex_t *preg, const char *exp, int cflags)
+hidden int jim_regcomp(regex_t *preg, const char *exp, int cflags)
 {
 	int scan;
 	int longest;
@@ -1161,7 +1170,7 @@ static int regrepeat(regex_t *preg, int p, int max);
 /*
  - regexec - match a regexp against a string
  */
-int jim_regexec(regex_t  *preg,  const  char *string, size_t nmatch, regmatch_t pmatch[], int eflags)
+hidden int jim_regexec(regex_t  *preg,  const  char *string, size_t nmatch, regmatch_t pmatch[], int eflags)
 {
 	const char *s;
 	int scan;
@@ -1916,7 +1925,7 @@ static const char *regprop( int op )
 }
 #endif /* JIM_BOOTSTRAP */
 
-size_t jim_regerror(int errcode, const regex_t *preg, char *errbuf,  size_t errbuf_size)
+hidden size_t jim_regerror(int errcode, const regex_t *preg, char *errbuf,  size_t errbuf_size)
 {
 	static const char *error_strings[] = {
 		"success",
@@ -1953,7 +1962,7 @@ size_t jim_regerror(int errcode, const regex_t *preg, char *errbuf,  size_t errb
 	return snprintf(errbuf, errbuf_size, "%s", err);
 }
 
-void jim_regfree(regex_t *preg)
+hidden void jim_regfree(regex_t *preg)
 {
 	free(preg->program);
 }