+++ CHANGELOG.md
@@ -12,7 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Use [Jim Tcl]
+- Use [Jim Tcl] and hide symbols `jim_regcomp(),` `jim_regerror()`,
+  `jim_regexec()`, and `jim_regfree()`
 
 ## [v23] - 2024-06-25